Utils Refactor | Game Schedule Result Conflict resolution
