+++ app/utils/config_utils.py
@@ -1,0 +1,17 @@
+import os
+import logging
+
+# Global Logging Configuration
+LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO").upper()
+LOG_FILE = os.getenv("LOG_FILE", "nba_data_module.log")
+
+logging.basicConfig(
+    level=LOG_LEVEL,
+    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
+    handlers=[logging.StreamHandler(), logging.FileHandler(LOG_FILE)],
+)
+
+logger = logging.getLogger(__name__)
+
+# Global Rate-Limiting for API Calls
+API_RATE_LIMIT = 0.6  # Seconds to sleep between API calls
