+++ ingest_data.py
@@ -1,11 +1,14 @@
 import logging
-from app.utils import (
+from app.utils.fetch.fetch_utils import (
     fetch_and_store_players,
     fetch_and_store_all_players_stats,
     fetch_and_store_current_rosters,
     fetch_and_store_leaguedashplayer_stats,
     fetch_and_store_leaguedashplayer_stats_for_current_season,
     fetch_and_store_player,
+)
+
+from app.utils.get.get_utils import (
     get_game_logs_for_all_players,
     get_game_logs_for_current_season,
     populate_schedule,
@@ -32,28 +35,28 @@ try:
     # Fetch teams, current rosters, and store to DB
     # Updates Team Rosters if exists
     # Run Daily
-    # fetch_and_store_current_rosters()
+    fetch_and_store_current_rosters()
     logging.info("Fetched and stored current rosters.")
 
     # Get Game Logs In X Seasons
     # Used to populate database
     # Run once or when specific season data is needed
     # get_game_logs_for_all_players(2015, 2024)
-    # logging.info("Fetched and stored game logs for all players in selected season(s)")
+    logging.info("Fetched and stored game logs for all players in selected season(s)")
 
     # Get Game Logs Current Season
     # Run Daily
-    # get_game_logs_for_current_season()
+    get_game_logs_for_current_season()
     logging.info("Fetched and stored game logs for all players in Current Season")
 
     # Fetch all career stats for all players and store in statistics table
     # Run Daily
-    #fetch_and_store_all_players_stats()
+    fetch_and_store_all_players_stats()
     logging.info("Fetched and stored all players' stats.")
 
     # Fetch the game schedule for the current season
     # Run Daily
-    #populate_schedule()
+    populate_schedule()
     logging.info("Populated the game schedule.")
 
     # Populate League Dash Player Stats for Last XXXX-XXXX Seasons
